Locations database on SQLite, IDs are long values
-Implemented the first parts of the SQL database.
-The Locations database now runs fully on SQLite, and other databases
are ready for use.
-Changed all of the IDs to long values for wider range and SQL
compatibility.
-Implemented search widget (instead of simple collapsable edittext) for
Locations (others will follow).

+++ PhotoTools/src/hu/bute/gb/onlab/PhotoTools/DeadlinesActivity.java
@@ -23,7 +23,7 @@ import com.slidingmenu.lib.SlidingMenu;
 
 public class DeadlinesActivity extends SherlockFragmentActivity {
 
-	public List<Integer> deadlinesOnView = new ArrayList<Integer>();
+	public List<Long> deadlinesOnView = new ArrayList<Long>();
 
 	private ViewGroup fragmentContainer_;
 	private FragmentManager fragmentManager_;
@@ -134,7 +134,7 @@ public class DeadlinesActivity extends SherlockFragmentActivity {
 			case 2:
 				if (resultCode == RESULT_OK) {
 					// The id of the location to add
-					Integer id = Integer.valueOf(data.getIntExtra("addedid", 0));
+					Long id = Long.valueOf(data.getIntExtra("addedid", 0));
 
 					deadlinesOnView.add(id);
 					String toAdd = data.getStringExtra("addedname");
